+++ packages/react-native-lazy-index/test/__fixtures__/BatchedBridge/package.json
@@ -1,6 +1,6 @@
 {
-  "experiences": [
-    "./module-01",
-    "./module-02"
-  ]
+  "experiences": {
+    "callable:Module-01": "./module-01",
+    "callable:Module-02": "./module-02"
+  }
 }
