--- packages/react-native-lazy-index/test/__fixtures__/MyOtherAwesomeApp/package.json
@@ -1,9 +1,0 @@
-{
-  "experiences": {
-    "callable:YetAnotherFeature": "@awesome-app/yet-another-feature",
-    "callable:YetAnotherFeatureLazy": "@awesome-app/yet-another-feature",
-    "callable:AnotherFeature": "@awesome-app/another-feature",
-    "SomeFeature": "@awesome-app/some-feature",
-    "FinalFeature": "@awesome-app/final-feature"
-  }
-}
