+++ packages/react-native-lazy-index/test/index.test.ts
@@ -2,7 +2,8 @@ import type { BabelFileResult } from "@babel/core";
 
 describe("react-native-lazy-index", () => {
   const babel = require("@babel/core");
-  const path = require("path");
+  const path = require("node:path");
+
   const currentWorkingDir = process.cwd();
 
   /**
@@ -11,11 +12,14 @@ describe("react-native-lazy-index", () => {
   function transformFixture(fixture: string): BabelFileResult | null {
     const workingDir = path.join(__dirname, "__fixtures__", fixture);
     process.chdir(workingDir);
-    return babel.transformFileSync("../../../src/index.js", {
-      cwd: workingDir,
-      filename: `${fixture}.js`,
-      plugins: ["codegen"],
-    });
+    return babel.transformSync(
+      `// @codegen\nmodule.exports = require("../../../src/index")();`,
+      {
+        cwd: workingDir,
+        filename: `${fixture}.js`,
+        plugins: ["codegen"],
+      }
+    );
   }
 
   afterEach(() => process.chdir(currentWorkingDir));
@@ -34,12 +38,6 @@ describe("react-native-lazy-index", () => {
 
   test("wraps registered components", () => {
     const result = transformFixture("MyAwesomeApp");
-    expect(result).toBeTruthy();
-    expect(result?.code).toMatchSnapshot();
-  });
-
-  test("wraps registered components using declared entry points", () => {
-    const result = transformFixture("MyOtherAwesomeApp");
     expect(result).toBeTruthy();
     expect(result?.code).toMatchSnapshot();
   });
